Remove pdffit2 and pdffit2server scripts.
Starting python in interactive mode as was done for the pdffit2
launcher is hard to do across all platforms.  pdffit2 launcher
has been converted to IPython extension diffpy.pdffit2.ipy_ext.

pdffit2server is unlikely to be used in the future.

Do not install any scripts with the diffpy.pdffit2 Python package.